pyproject.toml
[project]
name = "zatugaku"
version = "1.0.0"
description = "雑学動画ジェネレータ"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.8.1",
    "diffusers>=0.34.0",
    "gradio>=5.35.0",
    "llama-index-core>=0.12.46",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-vector-stores-chroma>=0.4.2",
    "mcp>=1.10.1",
    "moviepy>=2.2.1",
    "pillow>=11.3.0",
    "pyopenjtalk>=0.4.1",
    "scipy>=1.16.0",
    "sentencepiece>=0.2.0",
    "tomli>=2.2.1",
    "torchaudio>=2.7.1",
]
